third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/getTextureFormat-rgba16f.https.html

<!DOCTYPE html>
<title>Canvas2D WebGPU Transfer: getTextureFormat-rgba16float</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./canvas-test.js"></script>

<h1>Canvas2D WebGPU Transfer: getTextureFormat-rgba16f</h1>
<p>getTextureFormat() returns RGBA16F for a float16 context</p>

<script>

runCanvasTestsInWorker();

/** getTextureFormat() should return RGBA16F for a float16 context. */
canvasPromiseTest(async (canvas) => {
  const ctx = canvas.getContext('2d', {colorSpace: 'display-p3',
                                       pixelFormat: 'float16'});
  assert_equals(ctx.getTextureFormat(), 'rgba16float');
}, 'getTextureFormat() returns RGBA16F for a float16 context');

</script>
